+++ src/helper_functions.c
@@ -6,13 +6,13 @@
 /*   By: klukiano <klukiano@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/05 13:05:18 by klukiano          #+#    #+#             */
-/*   Updated: 2024/01/16 17:09:15 by klukiano         ###   ########.fr       */
+/*   Updated: 2024/01/25 18:28:50 by klukiano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../lib/libft/libft.h"
 
-int	free_n_0(int **int_array, char **str_array)
+int	free_n_0(long **int_array, char **str_array)
 {
 	int	i;
